Changed file structure to get rid of duplicate code for countdown
