+++ app/components/countdowns/bottom_countdown/BottomCountdownSection.tsx
@@ -1,26 +1,28 @@
 "use client";
 
-import { usebottomCountdownTimer } from "./BottomCountdownTimer";
+import { useCountdown } from "../hooks/useCountdown";
 import DigitGroup from "./DigitGroup";
-import RegisterButton from "../RegisterButton";
+import RegisterButton from "../../RegisterButton";
+
+const TARGET_DATE = new Date("2026-11-17T00:00:00").getTime();
 
 export default function BottomCountdownSection() {
-  const { days, hours, minutes } = usebottomCountdownTimer();
+  const { days, hours, minutes } = useCountdown(TARGET_DATE);
 
   return (
     <section className="min-h-screen bg-sen-blue flex items-center justify-center px-6">
-      {/* Countdown wrapper */}
+      {/* Container */}
       <div className="w-full max-w-[1400px] flex flex-col items-center text-center">
         {/* Heading */}
         <h2 className="text-white font-bold tracking-[0.25em] text-[clamp(1.75rem,3.5vw,2.5rem)] mb-[clamp(3rem,6vw,5rem)]">
           TIME LEFT TO REGISTER
         </h2>
 
-        {/* Digit groups for days, hours, and minutes */}
+        {/* Countdown digits */}
         <div className="flex justify-center items-start gap-[clamp(3rem,7vw,7rem)] w-full">
-          <DigitGroup value={days} digits={3} label="DAYS" />
-          <DigitGroup value={hours} digits={2} label="HOURS" />
-          <DigitGroup value={minutes} digits={2} label="MINUTES" />
+          <DigitGroup value={days ?? 0} digits={3} label="DAYS" />
+          <DigitGroup value={hours ?? 0} digits={2} label="HOURS" />
+          <DigitGroup value={minutes ?? 0} digits={2} label="MINUTES" />
         </div>
 
         {/* Registration button */}
