+++ app/components/countdowns/hooks/useCountdown.ts
@@ -2,20 +2,18 @@
 
 import { useEffect, useState } from "react";
 
-const TARGET_DATE = new Date("2026-11-17T00:00:00").getTime();
-
-export function usebottomCountdownTimer() {
-  const [timeLeft, setTimeLeft] = useState(TARGET_DATE - Date.now());
+export function useCountdown(targetDate: number) {
+  const [timeLeft, setTimeLeft] = useState<number>(
+    targetDate - Date.now()
+  );
 
   useEffect(() => {
-    // Update countdown every second
     const timer = setInterval(() => {
-      setTimeLeft(TARGET_DATE - Date.now());
+      setTimeLeft(targetDate - Date.now());
     }, 1000);
 
-    // Cleanup interval
     return () => clearInterval(timer);
-  }, []);
+  }, [targetDate]);
 
   const ms = Math.max(timeLeft, 0);
 
@@ -23,5 +21,6 @@ export function usebottomCountdownTimer() {
     days: Math.floor(ms / 1000 / 60 / 60 / 24),
     hours: Math.floor((ms / 1000 / 60 / 60) % 24),
     minutes: Math.floor((ms / 1000 / 60) % 60),
+    seconds: Math.floor((ms / 1000) % 60),
   };
 }
